working on listing available resources from the new dataset
