+++ scripts/Colonies.js
@@ -7,44 +7,44 @@ const availableResources = getAvailableResources()
 const mineralsAtFacility = getMineralAtFacility()
 const minerals = getMinerals()
 
-export const listAvailableResources = () => {
-    let html = `<ul>`
+// export const listAvailableResources = () => {
+//     let html = `<ul>`
 
-    const foundColony = colonies.find(
-        (colony) => {
-            for (const governor of governorsArray) {
-                return colony.id === governor.colonyId
-            }
-        }
-    )
+//     const foundColony = colonies.find(
+//         (colony) => {
+//             for (const governor of governorsArray) {
+//                 return colony.id === governor.colonyId
+//             }
+//         }
+//     )
 
-    const foundResources = availableResources.find(
-        (resource) => {
-            if (resource.colonyId === foundColony.id) {
-                return resource
-            }
-        }
-    )
+//     const foundResources = availableResources.find(
+//         (resource) => {
+//             if (resource.colonyId === foundColony.id) {
+//                 return resource
+//             }
+//         }
+//     )
 
-    const foundFacilityMinerals = mineralsAtFacility.find(
-        (facilityMineral) => {
-            if (foundResources.mineralId === facilityMineral.id) {
-                return facilityMineral
-            }
-        }
-    )
+//     const foundFacilityMinerals = mineralsAtFacility.find(
+//         (facilityMineral) => {
+//             if (foundResources.mineralId === facilityMineral.id) {
+//                 return facilityMineral
+//             }
+//         }
+//     )
 
-    const mineralName = minerals.find(
-        (mineral) => {
-            if (foundFacilityMinerals.mineralId === mineral.id) {
-                return mineral
-            }
-        }
-    )
+//     const mineralName = minerals.find(
+//         (mineral) => {
+//             if (foundFacilityMinerals.mineralId === mineral.id) {
+//                 return mineral
+//             }
+//         }
+//     )
 
-    html += `<li> ${mineralName.name} </li>`
+//     html += `<li> ${mineralName.name} </li>`
 
-    html += `</ul>`
-    return html
+//     html += `</ul>`
+//     return html
 
-}
+// }
